refactor: rename helper args to cli
